+++ main.go
@@ -1,0 +1,106 @@
+package main
+
+import (
+	"encoding/json"
+	"github.com/gorilla/mux"
+	"log"
+	"net/http"
+	"strconv"
+	"time"
+)
+
+type Note struct {
+	Title       string    `json:"title"`
+	Description string    `json:"description"`
+	CreatedAt   time.Time `json:"create_at"`
+}
+
+var noteStore = make(map[string]Note)
+
+var id int
+
+// GetNoteHandler - GET - /api/notes
+func GetNoteHandler(w http.ResponseWriter, r *http.Request) {
+	var notes []Note
+	for _, v := range noteStore {
+		notes = append(notes, v)
+	}
+	w.Header().Set("Content-Type", "application/json")
+	j, err := json.Marshal(notes)
+	if err != nil {
+		panic(err)
+	}
+	w.WriteHeader(http.StatusOK)
+	w.Write(j)
+}
+
+// PostNoteHandler - POST - /api/notes
+func PostNoteHandler(w http.ResponseWriter, r *http.Request) {
+	var note Note
+	err := json.NewDecoder(r.Body).Decode(&note) // Receive the json Ruquest, Decode that into Go object and Decode to the object Note
+	if err != nil {                              //This should not be in production
+		panic(err)
+	}
+	note.CreatedAt = time.Now()
+	id++
+	k := strconv.Itoa(id)
+	noteStore[k] = note
+	// send the Go object to json
+	w.Header().Set("Content-Type", "application/json")
+	j, err := json.Marshal(note)
+	if err != nil {
+		panic(err)
+	}
+	w.WriteHeader(http.StatusCreated)
+	w.Write(j)
+}
+
+// PutNoteHandler - PUT - /api/notes
+func PutNoteHandler(w http.ResponseWriter, r *http.Request) {
+	vars := mux.Vars(r)
+	k := vars["id"]
+	var noteUpdate Note
+	err := json.NewDecoder(r.Body).Decode(&noteUpdate)
+	if err != nil {
+		panic(err)
+	}
+	if note, ok := noteStore[k]; ok {
+		noteUpdate.CreatedAt = note.CreatedAt
+		delete(noteStore, k)
+		noteStore[k] = noteUpdate
+	} else {
+		log.Printf("id %s not found", k)
+	}
+	w.WriteHeader(http.StatusNoContent)
+}
+
+// DeleteNoteHandler - DELETE - /api/notes
+func DeleteNoteHandler(w http.ResponseWriter, r *http.Request) {
+	vars := mux.Vars(r)
+	k := vars["id"]
+	if _, ok := noteStore[k]; ok {
+		delete(noteStore, k)
+	} else {
+		log.Printf("id %s not found", k)
+	}
+	w.WriteHeader(http.StatusNoContent)
+}
+
+func main() {
+	r := mux.NewRouter().StrictSlash(false)
+	r.HandleFunc("/api/notes", GetNoteHandler).Methods("GET")
+	r.HandleFunc("/api/notes", PostNoteHandler).Methods("POST")
+	r.HandleFunc("/api/notes/{id}", PutNoteHandler).Methods("PUT")
+	r.HandleFunc("/api/notes/{id}", DeleteNoteHandler).Methods("DELETE")
+
+	server := &http.Server{
+		Addr:           ":8080",
+		Handler:        r,
+		ReadTimeout:    10 * time.Second,
+		WriteTimeout:   10 * time.Second,
+		MaxHeaderBytes: 1 << 20,
+	}
+
+	log.Println("Listening....localhost:8080")
+	server.ListenAndServe()
+}