+++ CHANGELOG.md
@@ -8,10 +8,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.0.1] - 2017-10-10
 ### Added
-- GITHUB REPO
-- README.md
-- .gitignore for osx and GOLang
-- CHANGELOG.md
-- Changelog link in README
+- [GITHUB](https://github.com/SadotCorts/GOLangAPIRest) **REPO**
+- [README.md](./README.md)
+- [.gitignore](./.gitignore) **for osx and GOLang**
+- [CHANGELOG]() **updated**
+- [CHANGELOG]() **link in README**
 ## Changed
-- README.md updated
+- [README.md](./README.md) **updated**
+
+## [0.0.2] - 2017-10-16
+### Added
+- [main.go](./main.go) **Main code**
+### Changed
+- [CHANGELOG]() **updated**
+- [.gitignore](./.gitignore) **updated, ``vim`` gitignore added.**
+- [README.md](./README.md) **updated**
+
+## [0.0.3] - 2017-10-17
+### Changed
+- [main.go](./main.go) **POST, PUT and DELETE Methods added**
+- [CHANGELOG]() **updated**